requirements.txt
bcrypt==4.0.1
boto3==1.26.163
botocore==1.29.163
cachetools==5.3.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
cryptography==41.0.1
dask==2023.6.0
facebook-scribe==2.0.post1
fluent-logger==0.10.0
fsspec==2023.6.0
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.0
httplib2==0.22.0
idna==3.4
importlib-metadata==6.6.0
jmespath==1.0.1
locket==1.0.0
msgpack==1.0.5
multipledispatch==0.6.0
numpy==1.24.3
oauthlib==3.2.2
packaging==23.1
pandas==2.0.2
paramiko==3.2.0
partd==1.4.0
protobuf==4.23.1
psutil==5.9.5
pyarrow==12.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyNaCl==1.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
redis==3.5.3
redis-py-cluster==2.1.3
requests==2.30.0
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.6.1
scratchpad-py==0.0.6
six==1.16.0
thrift==0.16.0
toolz==0.12.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
zipp==3.15.0
